docs: add documentation

+++ test.py
@@ -7,9 +7,13 @@ import matplotlib.pyplot as plt
 import omegaconf
 import stable_baselines3
 
-import helpers
 import src.helpers as helpers
 from evaluate import evaluate
+
+
+def _predict(controller, observation):
+    action, _ = controller.predict(observation, deterministic=True)
+    return action
 
 
 @hydra.main(version_base=None, config_path="configurations", config_name="test")
@@ -20,34 +24,36 @@ def test(configuration: omegaconf.DictConfig):
     print("[config_path]:", helpers.get_config_path())
     print("[output]:", hydra.core.hydra_config.HydraConfig.get().runtime.output_dir)
 
-    # output_directory = hydra.core.hydra_config.HydraConfig.get().runtime.output_dir
+    output_directory = configuration.get("output_path", None)
 
-    # helpers.save_configuration(configuration, "test")
+    if output_directory is not None and os.path.exists(output_directory):
+        target_configuration = helpers.load_configuration(
+            output_directory=output_directory,
+            expected_script="train",
+        )
 
-    # print(configuration.observations)
+        best_model_path = os.path.join(output_directory, "best_model.zip")
 
-    environment = hydra.utils.instantiate(
-        configuration.straight, render_mode="rgb_array"
-    )
+        controller, environment, _ = helpers.instantiate_configuration(
+            configuration=target_configuration,
+            output_directory=output_directory,
+            load_best_model=os.path.exists(best_model_path),
+            base_dir="./configurations",
+        )
 
-    for i in range(10):
-        environment.reset()
-        plt.imshow(environment.render())
-        plt.savefig(f"test_{i}.local.png")
-    # environment = stable_baselines3.common.monitor.Monitor(
-    #     env=environment,
-    #     filename=os.path.join(output_directory, "logs", "monitor.csv"),
-    # )
+        # observation, _ = environment.reset()
+        # action = _predict(controller, observation)
+        # print("[action]:", action)
 
-    # for wrapper in configuration.observations:
-    #     print("[wrapper]:")
-    #     print(wrapper)
-    #     environment = hydra.utils.instantiate(
-    #         wrapper,
-    #         environment=environment
-    #     )
+    else:
+        environment = hydra.utils.instantiate(
+            configuration.straight, render_mode="rgb_array"
+        )
 
-    # print(environment.observation_space)
+        for i in range(10):
+            environment.reset()
+            plt.imshow(environment.render())
+            plt.savefig(f"test_{i}.local.png")
 
 
 if __name__ == "__main__":
